+++ src/main/java/me/srrapero720/watermedia/api/loader/IEnvLoader.java
@@ -2,17 +2,20 @@ package me.srrapero720.watermedia.api.loader;
 
 /**
  * This is specially designed for modded environments.
- * In vanilla Java or... another type of modlauncher is not required
+ * In vanilla Java or another type of modlauncher is not required
  */
 public interface IEnvLoader {
     /**
-     * Checks for the current environments.
+     * Check if the current environment was a client
+     * this is important to prevent run on clients and load any client-side class like GL11
      * @return if was client-side
      */
     boolean client();
 
     /**
-     * Checks for any special trigger for development environment.
+     * Checks if the current environment was a development environment
+     * We cannot consider offline-mode as a dev env because _cracked users_<br>
+     * Is necessary search any special trigger like forge
      * @return if was a development env
      */
     boolean development();
@@ -28,6 +31,7 @@ public interface IEnvLoader {
     /**
      * Aggressive trigger for TLauncher
      * this was used by internal usages of WATERMeDIA.
+     * Please DO NOT bypass
      * @return if the current environment is using TLauncher
      */
     boolean tlauncher();