+++ src/main/java/me/srrapero720/watermedia/api/rendering/RenderAPI.java
@@ -1,0 +1,113 @@
+package me.srrapero720.watermedia.api.rendering;
+
+import me.lib720.watermod.reflect.ReflectTool;
+import me.srrapero720.watermedia.api.image.ImageRenderer;
+import org.lwjgl.BufferUtils;
+import org.lwjgl.opengl.GL11;
+import org.lwjgl.opengl.GL12;
+
+import java.awt.image.BufferedImage;
+import java.nio.ByteBuffer;
+import java.nio.IntBuffer;
+
+/**
+ * RenderApi is a tool class for OpenGL rendering compatible with all minecraft versions
+ */
+public class RenderAPI {
+
+    /**
+     * Created by CreativeMD
+     * Creates a new texture id based on a {@link BufferedImage} instance
+     * (used internally by {@link ImageRenderer}
+     * @param image image to process
+     * @param width buffer width (can be image width)
+     * @param height buffer height (can be image height)
+     * @return texture id for OpenGL
+     */
+    public static int applyBuffer(BufferedImage image, int width, int height) {
+        int[] pixels = new int[width * height];
+        image.getRGB(0, 0, width, height, pixels, 0, width);
+        boolean alpha = false;
+
+        if (image.getColorModel().hasAlpha()) for (int pixel : pixels)
+            if ((pixel >> 24 & 0xFF) < 0xFF) {
+                alpha = true;
+                break;
+            }
+
+        int bytesPerPixel = alpha ? 4 : 3;
+        ByteBuffer buffer = BufferUtils.createByteBuffer(width * height * bytesPerPixel);
+        for (int pixel : pixels) {
+            buffer.put((byte) ((pixel >> 16) & 0xFF)); // Red
+            buffer.put((byte) ((pixel >> 8) & 0xFF)); // Green
+            buffer.put((byte) (pixel & 0xFF)); // Blue
+            if (alpha) buffer.put((byte) ((pixel >> 24) & 0xFF)); // Alpha
+        }
+
+        /*
+         * FLIP method changes what class type returns in new JAVA versions, in runtime causes a JVM crash by that
+         * THIS EXECUTES {@link ByteBuffer#flip }
+         */
+        ReflectTool.invoke("flip", buffer.getClass(), buffer);
+
+        int textureID = GL11.glGenTextures(); //Generate texture ID
+        GL11.glBindTexture(GL11.GL_TEXTURE_2D, textureID); // Bind texture ID
+
+        //Setup wrap mode
+        GL11.glTexParameteri(GL11.GL_TEXTURE_2D, GL11.GL_TEXTURE_WRAP_S, GL12.GL_CLAMP_TO_EDGE);
+        GL11.glTexParameteri(GL11.GL_TEXTURE_2D, GL11.GL_TEXTURE_WRAP_T, GL12.GL_CLAMP_TO_EDGE);
+
+        //Setup texture scaling filtering
+        GL11.glTexParameteri(GL11.GL_TEXTURE_2D, GL11.GL_TEXTURE_MIN_FILTER, GL11.GL_LINEAR);
+        GL11.glTexParameteri(GL11.GL_TEXTURE_2D, GL11.GL_TEXTURE_MAG_FILTER, GL11.GL_LINEAR);
+
+        if (!alpha) GL11.glPixelStorei(GL11.GL_UNPACK_ALIGNMENT, GL11.GL_ONE);
+
+        // prevents random crash; when values are too high it causes a jvm crash, caused weird behavior when game is paused
+        GL11.glPixelStorei(GL11.GL_UNPACK_ROW_LENGTH, GL11.GL_ZERO);
+        GL11.glPixelStorei(GL11.GL_UNPACK_SKIP_PIXELS, GL11.GL_ZERO);
+        GL11.glPixelStorei(GL11.GL_UNPACK_SKIP_ROWS, GL11.GL_ZERO);
+
+        //Send texel data to OpenGL
+        GL11.glTexImage2D(GL11.GL_TEXTURE_2D, 0, alpha ? GL11.GL_RGBA8 : GL11.GL_RGB8, width, height, 0, alpha ? GL11.GL_RGBA : GL11.GL_RGB, GL11.GL_UNSIGNED_BYTE, buffer);
+
+        //Return the texture ID, so we can bind it later again
+        return textureID;
+    }
+
+    /**
+     * Process a buffer to be used in a OpenGL texture id
+     * @param videoBuffer IntBuffer to be processed
+     * @param glTexture texture ID from OpenGL
+     * @param videoWidth buffer width
+     * @param videoHeight buffer height
+     * @param firstFrame if was the first frame
+     */
+    public static void applyBuffer(IntBuffer videoBuffer, int glTexture, int videoWidth, int videoHeight, boolean firstFrame) {
+        GL11.glPixelStorei(GL11.GL_UNPACK_ROW_LENGTH, GL11.GL_ZERO);
+        GL11.glPixelStorei(GL11.GL_UNPACK_SKIP_PIXELS, GL11.GL_ZERO);
+        GL11.glPixelStorei(GL11.GL_UNPACK_SKIP_ROWS, GL11.GL_ZERO);
+
+        GL11.glBindTexture(GL11.GL_TEXTURE_2D, glTexture);
+        if (firstFrame) {GL11.glTexImage2D(GL11.GL_TEXTURE_2D, 0, GL11.GL_RGBA, videoWidth, videoHeight, 0, GL11.GL_RGBA, GL11.GL_UNSIGNED_BYTE, videoBuffer);
+        } else GL11.glTexSubImage2D(GL11.GL_TEXTURE_2D, 0, 0, 0, videoWidth, videoHeight, GL11.GL_RGBA, GL11.GL_UNSIGNED_BYTE, videoBuffer);
+    }
+
+    /**
+     * Process a buffer to be used in a OpenGL texture id
+     * @param videoBuffer ByteBuffer to be processed
+     * @param glTexture texture ID from OpenGL
+     * @param videoWidth buffer width
+     * @param videoHeight buffer height
+     * @param firstFrame if was the first frame
+     */
+    public static void applyBuffer(ByteBuffer videoBuffer, int glTexture, int videoWidth, int videoHeight, boolean firstFrame) {
+        GL11.glPixelStorei(GL11.GL_UNPACK_ROW_LENGTH, GL11.GL_ZERO);
+        GL11.glPixelStorei(GL11.GL_UNPACK_SKIP_PIXELS, GL11.GL_ZERO);
+        GL11.glPixelStorei(GL11.GL_UNPACK_SKIP_ROWS, GL11.GL_ZERO);
+
+        GL11.glBindTexture(GL11.GL_TEXTURE_2D, glTexture);
+        if (firstFrame) {GL11.glTexImage2D(GL11.GL_TEXTURE_2D, 0, GL11.GL_RGBA, videoWidth, videoHeight, 0, GL11.GL_RGBA, GL11.GL_UNSIGNED_BYTE, videoBuffer);
+        } else GL11.glTexSubImage2D(GL11.GL_TEXTURE_2D, 0, 0, 0, videoWidth, videoHeight, GL11.GL_RGBA, GL11.GL_UNSIGNED_BYTE, videoBuffer);
+    }
+}