+++ src/main/java/me/srrapero720/watermedia/api/rendering/RenderApi.java
@@ -1,8 +1,0 @@
-package me.srrapero720.watermedia.api.rendering;
-
-/**
- * RenderApi is a tool class for OpenGL rendering compatible with all minecraft versions
- */
-public class RenderApi {
-
-}