+++ src/main/java/me/srrapero720/watermedia/api/rendering/RenderGL.java
@@ -1,4 +1,0 @@
-package me.srrapero720.watermedia.api.rendering;
-
-public class RenderGL {
-}