+++ src/main/java/me/srrapero720/watermedia/api/rendering/RenderManager.java
@@ -1,5 +1,0 @@
-package me.srrapero720.watermedia.api.rendering;
-
-class RenderManager {
-
-}