+++ src/main/java/me/srrapero720/watermedia/api/rendering/manager/RenderState.java
@@ -1,0 +1,4 @@
+package me.srrapero720.watermedia.api.rendering.manager;
+
+public class RenderState {
+}